+++ tests/models/esm/test_modeling_esm.py
@@ -33,7 +33,6 @@ if is_torch_available():
         EsmEmbeddings,
         create_position_ids_from_input_ids,
     )
-    from transformers.models.esm.modeling_esmfold import EsmForProteinFolding
 
 
 # copied from tests.test_modeling_roberta
@@ -129,97 +128,6 @@ class EsmModelTester:
         model.eval()
         result = model(input_ids, attention_mask=input_mask, labels=token_labels)
         self.parent.assertEqual(result.logits.shape, (self.batch_size, self.seq_length, self.num_labels))
-
-    def prepare_config_and_inputs_for_common(self):
-        config_and_inputs = self.prepare_config_and_inputs()
-        (
-            config,
-            input_ids,
-            input_mask,
-            sequence_labels,
-            token_labels,
-            choice_labels,
-        ) = config_and_inputs
-        inputs_dict = {"input_ids": input_ids, "attention_mask": input_mask}
-        return config, inputs_dict
-
-
-class EsmFoldModelTester:
-    def __init__(
-        self,
-        parent,
-    ):
-        self.parent = parent
-        self.batch_size = 13
-        self.seq_length = 7
-        self.is_training = False
-        self.use_input_mask = True
-        self.use_token_type_ids = False
-        self.use_labels = False
-        self.vocab_size = 19
-        self.hidden_size = 32
-        self.num_hidden_layers = 5
-        self.num_attention_heads = 4
-        self.intermediate_size = 37
-        self.hidden_act = "gelu"
-        self.hidden_dropout_prob = 0.1
-        self.attention_probs_dropout_prob = 0.1
-        self.max_position_embeddings = 512
-        self.type_vocab_size = 16
-        self.type_sequence_label_size = 2
-        self.initializer_range = 0.02
-        self.num_labels = 3
-        self.num_choices = 4
-        self.scope = None
-
-    def prepare_config_and_inputs(self):
-        input_ids = ids_tensor([self.batch_size, self.seq_length], self.vocab_size)
-
-        input_mask = None
-        if self.use_input_mask:
-            input_mask = random_attention_mask([self.batch_size, self.seq_length])
-
-        sequence_labels = None
-        token_labels = None
-        choice_labels = None
-        if self.use_labels:
-            sequence_labels = ids_tensor([self.batch_size], self.type_sequence_label_size)
-            token_labels = ids_tensor([self.batch_size, self.seq_length], self.num_labels)
-            choice_labels = ids_tensor([self.batch_size], self.num_choices)
-
-        config = self.get_config()
-
-        return config, input_ids, input_mask, sequence_labels, token_labels, choice_labels
-
-    def get_config(self):
-        config = EsmConfig(
-            vocab_size=33,
-            hidden_size=self.hidden_size,
-            pad_token_id=1,
-            num_hidden_layers=self.num_hidden_layers,
-            num_attention_heads=self.num_attention_heads,
-            intermediate_size=self.intermediate_size,
-            hidden_act=self.hidden_act,
-            hidden_dropout_prob=self.hidden_dropout_prob,
-            attention_probs_dropout_prob=self.attention_probs_dropout_prob,
-            max_position_embeddings=self.max_position_embeddings,
-            type_vocab_size=self.type_vocab_size,
-            initializer_range=self.initializer_range,
-            is_folding_model=True,
-            esmfold_config={"trunk": {"num_blocks": 2}, "fp16_esm": False},
-        )
-        return config
-
-    def create_and_check_model(self, config, input_ids, input_mask, sequence_labels, token_labels, choice_labels):
-        model = EsmForProteinFolding(config=config).float()
-        model.to(torch_device)
-        model.eval()
-        result = model(input_ids, attention_mask=input_mask)
-        result = model(input_ids)
-        result = model(input_ids)
-
-        self.parent.assertEqual(result.positions.shape, (8, self.batch_size, self.seq_length, 14, 3))
-        self.parent.assertEqual(result.angles.shape, (8, self.batch_size, self.seq_length, 7, 2))
 
     def prepare_config_and_inputs_for_common(self):
         config_and_inputs = self.prepare_config_and_inputs()
@@ -341,102 +249,6 @@ class EsmModelTest(ModelTesterMixin, unittest.TestCase):
 
 
 @require_torch
-class EsmFoldModelTest(ModelTesterMixin, unittest.TestCase):
-
-    test_mismatched_shapes = False
-
-    all_model_classes = (EsmForProteinFolding,) if is_torch_available() else ()
-    all_generative_model_classes = ()
-    test_sequence_classification_problem_types = False
-
-    def setUp(self):
-        self.model_tester = EsmFoldModelTester(self)
-        self.config_tester = ConfigTester(self, config_class=EsmConfig, hidden_size=37)
-
-    def test_config(self):
-        self.config_tester.run_common_tests()
-
-    def test_model(self):
-        config_and_inputs = self.model_tester.prepare_config_and_inputs()
-        self.model_tester.create_and_check_model(*config_and_inputs)
-
-    @unittest.skip("Does not support attention outputs")
-    def test_attention_outputs(self):
-        pass
-
-    @unittest.skip
-    def test_correct_missing_keys(self):
-        pass
-
-    @unittest.skip("Esm does not support embedding resizing")
-    def test_resize_embeddings_untied(self):
-        pass
-
-    @unittest.skip("Esm does not support embedding resizing")
-    def test_resize_tokens_embeddings(self):
-        pass
-
-    @unittest.skip("ESMFold does not support passing input embeds!")
-    def test_inputs_embeds(self):
-        pass
-
-    @unittest.skip("ESMFold does not support head pruning.")
-    def test_head_pruning(self):
-        pass
-
-    @unittest.skip("ESMFold does not support head pruning.")
-    def test_head_pruning_integration(self):
-        pass
-
-    @unittest.skip("ESMFold does not support head pruning.")
-    def test_head_pruning_save_load_from_config_init(self):
-        pass
-
-    @unittest.skip("ESMFold does not support head pruning.")
-    def test_head_pruning_save_load_from_pretrained(self):
-        pass
-
-    @unittest.skip("ESMFold does not support head pruning.")
-    def test_headmasking(self):
-        pass
-
-    @unittest.skip("ESMFold does not output hidden states in the normal way.")
-    def test_hidden_states_output(self):
-        pass
-
-    @unittest.skip("ESMfold does not output hidden states in the normal way.")
-    def test_retain_grad_hidden_states_attentions(self):
-        pass
-
-    @unittest.skip("ESMFold only has one output format.")
-    def test_model_outputs_equivalence(self):
-        pass
-
-    @unittest.skip("This test doesn't work for ESMFold and doesn't test core functionality")
-    def test_save_load_fast_init_from_base(self):
-        pass
-
-    @unittest.skip("ESMFold does not support input chunking.")
-    def test_feed_forward_chunking(self):
-        pass
-
-    @unittest.skip("ESMFold doesn't respect you and it certainly doesn't respect your initialization arguments.")
-    def test_initialization(self):
-        pass
-
-    @unittest.skip("ESMFold doesn't support torchscript compilation.")
-    def test_torchscript_output_attentions(self):
-        pass
-
-    @unittest.skip("ESMFold doesn't support torchscript compilation.")
-    def test_torchscript_output_hidden_state(self):
-        pass
-
-    @unittest.skip("ESMFold doesn't support torchscript compilation.")
-    def test_torchscript_simple(self):
-        pass
-
-@require_torch
 class EsmModelIntegrationTest(TestCasePlus):
     @slow
     def test_inference_masked_lm(self):
@@ -469,12 +281,3 @@ class EsmModelIntegrationTest(TestCasePlus):
                 [[[0.1444, 0.5413, 0.3248], [0.3034, 0.0053, 0.3108], [0.3228, -0.2499, 0.3415]]]
             )
             self.assertTrue(torch.allclose(output[:, :3, :3], expected_slice, atol=1e-4))
-
-    @slow
-    def test_inference_protein_folding(self):
-        model = EsmForProteinFolding.from_pretrained("Rocketknight1/esmfold_v1").float()
-        model.eval()
-        input_ids = torch.tensor([[0, 6, 4, 13, 5, 4, 16, 12, 11, 7, 2]])
-        position_outputs = model(input_ids)["positions"]
-        expected_slice = torch.tensor([2.5828, 0.7993, -10.9334], dtype=torch.float32)
-        self.assertTrue(torch.allclose(position_outputs[0, 0, 0, 0], expected_slice, atol=1e-4))
