+++ utils/check_inits.py
@@ -268,7 +268,6 @@ def get_transformers_submodules():
 IGNORE_SUBMODULES = [
     "convert_pytorch_checkpoint_to_tf2",
     "modeling_flax_pytorch_utils",
-    "models.esm.openfold_np",
     "models.esm.openfold_utils",
 ]
