+++ utils/check_repo.py
@@ -443,10 +443,7 @@ def check_all_models_are_tested():
     test_files = get_model_test_files()
     failures = []
     for module in modules:
-        module_name = module.__name__
-        if "esmfold" in module_name:
-            module_name = module_name.replace("esmfold", "esm")
-        test_file = [file for file in test_files if f"test_{module_name.split('.')[-1]}.py" in file]
+        test_file = [file for file in test_files if f"test_{module.__name__.split('.')[-1]}.py" in file]
         if len(test_file) == 0:
             failures.append(f"{module.__name__} does not have its corresponding test file {test_file}.")
         elif len(test_file) > 1:
